+++ tests/test_move_command.rs
@@ -371,8 +371,10 @@ fn test_move_description() {
 JJ: Enter a description for the combined commit.
 JJ: Description from the destination commit:
 destination
+
 JJ: Description from the source commit:
 source
+
 JJ: Lines starting with "JJ: " (like this one) will be removed.
 "#,
     )
@@ -380,6 +382,7 @@ JJ: Lines starting with "JJ: " (like this one) will be removed.
     test_env.jj_cmd_success(&repo_path, &["move", "--to", "@-"]);
     insta::assert_snapshot!(get_description(&test_env, &repo_path, "@-"), @r###"
     destination
+
     source
     "###);