+++ tests/test_obslog_command.rs
@@ -34,13 +34,13 @@ fn test_obslog_with_or_without_diff() {
 
     let stdout = get_log_output(&test_env, &repo_path, &["obslog"]);
     insta::assert_snapshot!(stdout, @r###"
-    @  test.user@example.com 2001-02-03 04:05:10.000 +07:00 1daafc17fefb
+    @  test.user@example.com 2001-02-03 04:05:10.000 +07:00 66b42ad36073
     | my description
-    o  test.user@example.com 2001-02-03 04:05:09.000 +07:00 813918f7b4e6 conflict
+    o  test.user@example.com 2001-02-03 04:05:09.000 +07:00 af536e5af67e conflict
     | my description
-    o  test.user@example.com 2001-02-03 04:05:09.000 +07:00 8f02f5470c55
+    o  test.user@example.com 2001-02-03 04:05:09.000 +07:00 6fbba7bcb590
     | my description
-    o  test.user@example.com 2001-02-03 04:05:08.000 +07:00 c8ceb219336b
+    o  test.user@example.com 2001-02-03 04:05:08.000 +07:00 eac0d0dae082
       my description
     "###);
 
@@ -48,43 +48,43 @@ fn test_obslog_with_or_without_diff() {
     // (even even though it resulted in a conflict).
     let stdout = get_log_output(&test_env, &repo_path, &["obslog", "-p"]);
     insta::assert_snapshot!(stdout, @r###"
-    @  test.user@example.com 2001-02-03 04:05:10.000 +07:00 1daafc17fefb
+    @  test.user@example.com 2001-02-03 04:05:10.000 +07:00 66b42ad36073
     | my description
     | Resolved conflict in file1:
     |    1    1: <<<<<<<resolved
     |    2     : %%%%%%%
     |    3     : +bar
     |    4     : >>>>>>>
-    o  test.user@example.com 2001-02-03 04:05:09.000 +07:00 813918f7b4e6 conflict
+    o  test.user@example.com 2001-02-03 04:05:09.000 +07:00 af536e5af67e conflict
     | my description
-    o  test.user@example.com 2001-02-03 04:05:09.000 +07:00 8f02f5470c55
+    o  test.user@example.com 2001-02-03 04:05:09.000 +07:00 6fbba7bcb590
     | my description
     | Modified regular file file1:
     |    1    1: foo
     |         2: bar
     | Added regular file file2:
     |         1: foo
-    o  test.user@example.com 2001-02-03 04:05:08.000 +07:00 c8ceb219336b
+    o  test.user@example.com 2001-02-03 04:05:08.000 +07:00 eac0d0dae082
       my description
     "###);
 
     // Test `--no-graph`
     let stdout = get_log_output(&test_env, &repo_path, &["obslog", "--no-graph"]);
     insta::assert_snapshot!(stdout, @r###"
-     test.user@example.com 2001-02-03 04:05:10.000 +07:00 1daafc17fefb
+     test.user@example.com 2001-02-03 04:05:10.000 +07:00 66b42ad36073
     my description
-     test.user@example.com 2001-02-03 04:05:09.000 +07:00 813918f7b4e6 conflict
+     test.user@example.com 2001-02-03 04:05:09.000 +07:00 af536e5af67e conflict
     my description
-     test.user@example.com 2001-02-03 04:05:09.000 +07:00 8f02f5470c55
+     test.user@example.com 2001-02-03 04:05:09.000 +07:00 6fbba7bcb590
     my description
-     test.user@example.com 2001-02-03 04:05:08.000 +07:00 c8ceb219336b
+     test.user@example.com 2001-02-03 04:05:08.000 +07:00 eac0d0dae082
     my description
     "###);
 
     // Test `--git` format, and that it implies `-p`
     let stdout = get_log_output(&test_env, &repo_path, &["obslog", "--no-graph", "--git"]);
     insta::assert_snapshot!(stdout, @r###"
-     test.user@example.com 2001-02-03 04:05:10.000 +07:00 1daafc17fefb
+     test.user@example.com 2001-02-03 04:05:10.000 +07:00 66b42ad36073
     my description
     diff --git a/file1 b/file1
     index e155302a24...2ab19ae607 100644
@@ -96,9 +96,9 @@ fn test_obslog_with_or_without_diff() {
     -+bar
     ->>>>>>>
     +resolved
-     test.user@example.com 2001-02-03 04:05:09.000 +07:00 813918f7b4e6 conflict
+     test.user@example.com 2001-02-03 04:05:09.000 +07:00 af536e5af67e conflict
     my description
-     test.user@example.com 2001-02-03 04:05:09.000 +07:00 8f02f5470c55
+     test.user@example.com 2001-02-03 04:05:09.000 +07:00 6fbba7bcb590
     my description
     diff --git a/file1 b/file1
     index 257cc5642c...3bd1f0e297 100644
@@ -114,7 +114,7 @@ fn test_obslog_with_or_without_diff() {
     +++ b/file2
     @@ -1,0 +1,1 @@
     +foo
-     test.user@example.com 2001-02-03 04:05:08.000 +07:00 c8ceb219336b
+     test.user@example.com 2001-02-03 04:05:08.000 +07:00 eac0d0dae082
     my description
     "###);
 }
@@ -141,20 +141,20 @@ fn test_obslog_squash() {
     | | Modified regular file file1:
     | |    1    1: foo
     | |         2: bar
-    o |  test.user@example.com 2001-02-03 04:05:09.000 +07:00 803a7299cb1a
+    o |  test.user@example.com 2001-02-03 04:05:09.000 +07:00 9764e503e1a9
     | | first
     | | Added regular file file1:
     | |         1: foo
-    o |  test.user@example.com 2001-02-03 04:05:08.000 +07:00 85a1e2839620
+    o |  test.user@example.com 2001-02-03 04:05:08.000 +07:00 69542c1984c1
     | | first
     o |  test.user@example.com 2001-02-03 04:05:07.000 +07:00 230dd059e1b0
      /  (no description set)
-    o  test.user@example.com 2001-02-03 04:05:10.000 +07:00 69231a40d60d
+    o  test.user@example.com 2001-02-03 04:05:10.000 +07:00 f09a38899f2b
     | second
     | Modified regular file file1:
     |    1    1: foo
     |         2: bar
-    o  test.user@example.com 2001-02-03 04:05:09.000 +07:00 b567edda97ab
+    o  test.user@example.com 2001-02-03 04:05:09.000 +07:00 579965369703
       second
     "###);
 }