+++ tests/test_rebase_command.rs
@@ -75,13 +75,13 @@ fn test_rebase_invalid() {
     // Rebase onto descendant with -r
     let stderr = test_env.jj_cmd_failure(&repo_path, &["rebase", "-r", "a", "-d", "b"]);
     insta::assert_snapshot!(stderr, @r###"
-    Error: Cannot rebase 873140c1fed9 onto descendant ad05f5d1407c
+    Error: Cannot rebase 2443ea76b0b1 onto descendant 1394f625cbbd
     "###);
 
     // Rebase onto descendant with -s
     let stderr = test_env.jj_cmd_failure(&repo_path, &["rebase", "-s", "a", "-d", "b"]);
     insta::assert_snapshot!(stderr, @r###"
-    Error: Cannot rebase 873140c1fed9 onto descendant ad05f5d1407c
+    Error: Cannot rebase 2443ea76b0b1 onto descendant 1394f625cbbd
     "###);
 }
 
@@ -150,7 +150,7 @@ fn test_rebase_branch_with_merge() {
     let stdout = test_env.jj_cmd_success(&repo_path, &["rebase", "-b", "d", "-d", "b"]);
     insta::assert_snapshot!(stdout, @r###"
     Rebased 3 commits
-    Working copy now at: b2674fa494af e
+    Working copy now at: 391c91a7defa e
     Added 1 files, modified 0 files, removed 0 files
     "###);
     insta::assert_snapshot!(get_log_output(&test_env, &repo_path), @r###"
@@ -166,7 +166,7 @@ fn test_rebase_branch_with_merge() {
     let stdout = test_env.jj_cmd_success(&repo_path, &["rebase", "-d", "b"]);
     insta::assert_snapshot!(stdout, @r###"
     Rebased 3 commits
-    Working copy now at: fef1da569696 e
+    Working copy now at: 040ae3a6d358 e
     Added 1 files, modified 0 files, removed 0 files
     "###);
     insta::assert_snapshot!(get_log_output(&test_env, &repo_path), @r###"
@@ -209,7 +209,7 @@ fn test_rebase_single_revision() {
     let stdout = test_env.jj_cmd_success(&repo_path, &["rebase", "-r", "b", "-d", "a"]);
     insta::assert_snapshot!(stdout, @r###"
     Also rebased 2 descendant commits onto parent of rebased commit
-    Working copy now at: ed4d09bb181f d
+    Working copy now at: 7e15b97a447f d
     Added 0 files, modified 0 files, removed 1 files
     "###);
     insta::assert_snapshot!(get_log_output(&test_env, &repo_path), @r###"
@@ -227,7 +227,7 @@ fn test_rebase_single_revision() {
     let stdout = test_env.jj_cmd_success(&repo_path, &["rebase", "-r", "c", "-d", "root"]);
     insta::assert_snapshot!(stdout, @r###"
     Also rebased 1 descendant commits onto parent of rebased commit
-    Working copy now at: 59a16d87a26f d
+    Working copy now at: bf87078f4560 d
     Added 0 files, modified 0 files, removed 1 files
     "###);
     insta::assert_snapshot!(get_log_output(&test_env, &repo_path), @r###"
@@ -269,7 +269,7 @@ fn test_rebase_single_revision_merge_parent() {
     let stdout = test_env.jj_cmd_success(&repo_path, &["rebase", "-r", "c", "-d", "a"]);
     insta::assert_snapshot!(stdout, @r###"
     Also rebased 1 descendant commits onto parent of rebased commit
-    Working copy now at: a4fccbb7582d d
+    Working copy now at: c62d0789e7c7 d
     Added 0 files, modified 0 files, removed 1 files
     "###);
     insta::assert_snapshot!(get_log_output(&test_env, &repo_path), @r###"
@@ -345,7 +345,7 @@ fn test_rebase_with_descendants() {
     let stdout = test_env.jj_cmd_success(&repo_path, &["rebase", "-s", "b", "-d", "a"]);
     insta::assert_snapshot!(stdout, @r###"
     Rebased 3 commits
-    Working copy now at: 9afba1135175 d
+    Working copy now at: 309336ffce22 d
     "###);
     insta::assert_snapshot!(get_log_output(&test_env, &repo_path), @r###"
     @ d