+++ tests/test_templater.rs
@@ -66,12 +66,12 @@ fn test_templater_branches() {
         &["log", "-T", r#"commit_id.short() " " branches"#],
     );
     insta::assert_snapshot!(output, @r###"
-    o 48e0b6c42296 branch3?
-    | @ 092b2e0283a9 branch2* new-branch
-    | | o f4a739b1677f branch1*
+    o b1bb3766d584 branch3?
+    | @ a5b4d15489cc branch2* new-branch
+    | | o 21c33875443e branch1*
     | |/  
     |/|   
-    | o 752dad8b1718 branch2@origin
+    | o 8476341eb395 branch2@origin
     |/  
     o 000000000000 
     "###);